Adapted engine to redone timers

+++ src/engine/ecs/system.hpp
@@ -1,12 +1,14 @@
 #ifndef __CHESTNUT_SYSTEM_H__
 #define __CHESTNUT_SYSTEM_H__
 
+#include <cstdint>
+
 namespace chestnut
 {
     class ISystem
     {
     public:
-        virtual void update( float deltaTime ) = 0;
+        virtual void update( uint32_t deltaTime ) = 0;
         virtual ~ISystem() {}
     };
