+++ src/engine/ecs/systems/timer_system.hpp
@@ -18,7 +18,7 @@ namespace chestnut
     public:
         void submitBatch( CComponentBatch *batch ) override;
         void clearBatches() override;
-        void update( float deltaTime ) override;
+        void update( uint32_t deltaTime ) override;
     };
 
 } // namespace chestnut
